pyproject.toml
[tool.poetry]
name = "cubesolver"
version = "0.1.0"
description = ""
authors = ["Naoki Pross <np@0hm.ch>", "Keita Pross"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9"
pygame = "^2.4.0"
pyopengl = "^3.1.7"
numpy = "^1.25.0"
scikit-image = "^0.21.0"
opencv-python = "^4.7.0.72"
pyopengl-accelerate = "^3.1.7"
ipython = "^8.14.0"
numpy-quaternion = "^2022.4.3"
matplotlib = "^3.7.1"
pyrr = "^0.10.3"


[tool.poetry.group.dev.dependencies]
notebook = "^6.5.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
